requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
frozenlist==1.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
pandas==2.2.3
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
ta==0.11.0
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
